+++ src/apycfixedarray.h
@@ -93,7 +93,7 @@ public:
     APyCFixedArray() = delete;
 
     explicit APyCFixedArray(
-        const nb::typed<nb::sequence, nb::any>& bit_pattern_sequence,
+        const nb::typed<nb::iterable, nb::any>& bit_pattern_sequence,
         std::optional<int> int_bits = std::nullopt,
         std::optional<int> frac_bits = std::nullopt,
         std::optional<int> bits = std::nullopt
@@ -252,7 +252,7 @@ public:
     //! Create an `APyCFixedArray` tensor object initialized with values from a sequence
     //! of `complex`
     static APyCFixedArray from_complex(
-        const nb::typed<nb::sequence, nb::any>& cplx_seq,
+        const nb::typed<nb::iterable, nb::any>& cplx_seq,
         std::optional<int> int_bits = std::nullopt,
         std::optional<int> frac_bits = std::nullopt,
         std::optional<int> bits = std::nullopt
@@ -261,7 +261,7 @@ public:
     //! Create an `APyCFixedArray` tensor object initialized with values from a sequence
     //! of numbers
     static APyCFixedArray from_numbers(
-        const nb::typed<nb::sequence, nb::any>& number_seq,
+        const nb::typed<nb::iterable, nb::any>& number_seq,
         std::optional<int> int_bits = std::nullopt,
         std::optional<int> frac_bits = std::nullopt,
         std::optional<int> bits = std::nullopt